% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{collect_stability_data}
\alias{collect_stability_data}
\title{Collect Stability Data for Multiple Packages}
\usage{
collect_stability_data(pkgs, releases = NULL, months = NULL)
}
\arguments{
\item{pkgs}{A vector of package names to collect stability data for.}

\item{releases}{An integer indicating the number of releases to collect
stability data for. For example, \code{releases = 10} will return stability
data for the last 10 releases of the package. Default is NULL, which means
the function will return data for all releases.}

\item{months}{An integer indicating the number of months back to collect
stability data for.  For example, \code{months = 24} will collect
stability data for the previous 2 years.  Default is NULL, meaning there is
no limitation on the number of release months, and the function will collect
data from all releases.}
}
\description{
The \strong{collect_stability_data} function function
}
\seealso{
Other stability: 
\code{\link{get_stability_data}()},
\code{\link{get_stability_score}()},
\code{\link{print.pdiff_score}()}
}
\concept{stability}
