% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_diff}
\alias{get_diff}
\title{Get a Package Difference Object}
\usage{
get_diff(pkgname, v1 = "current", v2 = "latest")
}
\arguments{
\item{pkgname}{The package name.}

\item{v1}{The earlier package version.}

\item{v2}{The later package version.}
}
\description{
Get a Package Difference Object
}
\seealso{
Other pdiff: 
\code{\link{print.pdiff}()},
\code{\link{view_details}()}
}
\concept{pdiff}
