% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_installed_packages}
\alias{get_installed_packages}
\title{Get Installed Packages and Versions}
\usage{
get_installed_packages()
}
\value{
A data frame containing each installed
package and its version number.
}
\description{
A utility function to return the installed packages
in the current repository and their
versions numbers.
}
\examples{
# Get all packages in repo
res <- get_installed_packages()

# View top 10 results
res[1:12, ]
#        Package Version
# 1        abind   1.4-8
# 2         ards   0.1.1
# 3      askpass   1.2.0
# 4   assertthat   0.2.1
# 5    backports   1.5.0
# 6    base64enc   0.1-3
# 7         bigD   0.3.0
# 8  BiocManager 1.30.23
# 9          bit   4.0.5
# 10       bit64   4.0.5
}
