% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{get_stability_data}
\alias{get_stability_data}
\title{Collect Stability Data for a Package}
\usage{
get_stability_data(pkgs, releases = NULL, months = NULL)
}
\arguments{
\item{pkgs}{A character vector of one or more package names.}

\item{releases}{An integer indicating the number of releases to collect
stability data for. For example, \code{releases = 10} will return stability
data for the last 10 releases of the package. Default is NULL, which means
the function will return data for all releases.}

\item{months}{An integer indicating the number of months back to collect
stability data for.  For example, \code{months = 24} will collect
stability data for the previous 2 years.  Default is NULL, meaning there is
no limitation on the number of release months, and the function will collect
data from all releases.}
}
\value{
A data frame of stability information for the targeted packages.
The data frame will contain one row for each release.  Each row will
contains comparison information against the prior release.  The
columns will be organized as follows:
\itemize{
\item {\strong{Version}: The version number of the release.
}
\item {\strong{FileName}: The file name of the package.
}
\item {\strong{Release}: The release date of the package.
}
\item {\strong{Size}: The size of the package file on disk.
}
\item {\strong{AF}: The number of added functions.
}
\item {\strong{AP}: The number of added parameters.
}
\item {\strong{RF}: The number of removed functions.
}
\item {\strong{RP}: The number of removed parameters.
}
\item {\strong{BC}: Whether the release had any breaking changes.
Breaking changes are removed functions or removed parameters.  Values
are zero or one.  The value one means the release had breaking changes.
}
\item {\strong{TF}: The total number of functions in the package.
}

}
}
\description{
Collect Stability Data for a Package
}
\seealso{
Other stability: 
\code{\link{get_stability_score}()},
\code{\link{print.pdiff_score}()}
}
\concept{stability}
