% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{get_stability_score}
\alias{get_stability_score}
\title{Calculate Stability Score for a Package}
\usage{
get_stability_score(pkgname, releases = NULL, months = NULL, source = "github")
}
\arguments{
\item{pkgname}{The name of the package.}

\item{releases}{An integer indicating the number of releases to collect
stability data for. For example, \code{releases = 10} will return stability
data for the last 10 releases of the package. Default is NULL, which means
the function will return data for all releases.}

\item{months}{An integer indicating the number of months back to collect
stability data for.  For example, \code{months = 24} will collect
stability data for the previous 2 years.  Default is NULL, meaning there is
no limitation on the number of release months, and the function will collect
data from all releases.}

\item{source}{The location to retrieve stability data.  Valid choices are
"github", or "cran".  The default is "github".}
}
\value{
An stability score object of class "pdiff_score".  The object
contains the stability score, plus other useful information such as
the release and version ranges, the number of releases, and number of
breaking releases.
}
\description{
The \code{get_stablity_score} function calculates a stability score for
a specified package.  The score is calculated as the percentage of
releases with no breaking changes.  A breaking change is defined as
either a removed function or removed function parameter. The object
also returns the raw data from which the stability score is calculated.
This data can be useful for review or even custom analytics.  For a
description of the stability data frame, see \code{\link{get_stability_data}}.
}
\seealso{
Other stability: 
\code{\link{get_stability_data}()},
\code{\link{print.pdiff_score}()}
}
\concept{stability}
