% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{print.pfunction}
\alias{print.pfunction}
\title{Print a Package Function Object}
\usage{
\method{print}{pfunction}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The package function to print.}

\item{...}{Follow-on parameters to the print function.}

\item{verbose}{Whether to print in summary or list-style.}
}
\description{
Print a Package Function Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
