% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_breakages}
\alias{report_breakages}
\title{Generate a Report of Upgrade Breakages}
\usage{
report_breakages(v1pkgs = "current", v2pkgs = "latest")
}
\arguments{
\item{v1pkgs}{A data frame that identifies the source packages
and versions.}

\item{v2pkgs}{A data frame that identifies the target packages
and versions.  By default,}
}
\description{
Generate a Report of Upgrade Breakages
}
\seealso{
Other reports: 
\code{\link{report_stability}()}
}
\concept{reports}
