% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altutils.R
\name{scanner}
\alias{scanner}
\title{Scan List of Text Files}
\usage{
scanner(files, full.names = FALSE)
}
\arguments{
\item{files}{the names of files to read.}

\item{full.names}{Whether to use full file names.}
}
\description{
Store text file as character string
}
\keyword{internal}
