% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{view_details}
\alias{view_details}
\title{View Package Difference Details}
\usage{
view_details(diff, docs = TRUE)
}
\arguments{
\item{diff}{A package difference object of class "pdiff".  This object is
returned from \code{\link{pkg_diff}}.}

\item{docs}{Whether to include the function documentation
in the viewer comparison.  Default is TRUE.}
}
\value{
A package version comparison is displyed in the RStudio viewer.
}
\description{
The \strong{view_details} function displays package differences
in the RStudio viewer, so they can be examined in detail. Package information
is taken from CRAN.
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()},
\code{\link{print.pdiff}()}
}
\concept{pdiff}
