% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{pkg_cache}
\alias{pkg_cache}
\title{Returns Information on the Package Cache}
\usage{
pkg_cache(pkgs = NULL)
}
\arguments{
\item{pkgs}{A vector of package names to retrieve cache information about.
Default is NULL, which means to return all packages in the cache.}
}
\value{
An data frame showing the package name and latest package
version stored in the cache.  If the package is not stored in the cache,
the package version will be NA.
}
\description{
The \code{pkg_cache} function queries the package cache, and
returns information on whether or not the package is part of the cache.
The function also returns the last version of the package cached.
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
