% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{pkg_cache}
\alias{pkg_cache}
\title{Returns Information on the Package Cache}
\usage{
pkg_cache(pkgs = NULL)
}
\arguments{
\item{pkgs}{A vector of package names to retrieve cache information about.
Default is NULL, which means to return all packages in the cache.}
}
\value{
An data frame showing the package name and latest package
version stored in the cache.  If the package is not stored in the cache,
the package version will be NA.
}
\description{
The \code{pkg_cache} function queries the package cache, and
returns information on which packages are included in the cache.
The function also returns the last version of the package cached.
}
\details{
The \strong{pkgdiff} cache is used to speed up \strong{pkgdiff} functions.
Information about each package is pre-processed and stored in the cache
on Github.  The functions then retrieve this pre-processed information
instead of pulling packages down from CRAN.  To learn more about the cache,
refer to \code{vignette('Package Cache')}.
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
