% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{pkg_diff}
\alias{pkg_diff}
\title{Get a Package Difference Object}
\usage{
pkg_diff(pkgname, v1 = "current", v2 = "latest")
}
\arguments{
\item{pkgname}{The package name.}

\item{v1}{The earlier package version.  Default is "current", which
means the function will look up the currently installed version.}

\item{v2}{The later package version.  The default is "latest", which is the
latest version of the package available on CRAN.}
}
\description{
Get a Package Difference Object
}
\seealso{
Other pdiff: 
\code{\link{print.pdiff}()},
\code{\link{view_details}()}
}
\concept{pdiff}
