% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{pkg_info}
\alias{pkg_info}
\title{Get Information for a Package}
\usage{
pkg_info(pkg, ver = "current", cache = TRUE)
}
\arguments{
\item{pkg}{The package name.}

\item{ver}{The package version. Special values are "current" and
"latest".  The value "current" is the current version of the package
running on the machine.  The value "latest" is the latest version
of the package from CRAN.}

\item{cache}{Whether to retrieve the info from the github cache, or
from CRAN.  If TRUE, the function will first search the cache, and
return the info if available.  If the info is not available in the
github cache, or the cache parameter is set to FALSE, the info will be
retrieved from CRAN.}
}
\description{
Combines information related to a package. Objects of this class are
used in several other functions in the \strong{pkgdiff} system.  This
class also makes a compact storage format for pre-processed package
information.
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_repo}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
