% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgrepo.R
\name{pkg_repo}
\alias{pkg_repo}
\title{A Package Repository}
\usage{
pkg_repo(pkgs = NULL, ver = "current", libpaths = NULL)
}
\arguments{
\item{pkgs}{A vector of package names used to subset the repository list.
Default is NULL, which means all packages in the repository will be returned.}

\item{ver}{The R version of the repository. Default is "current",
meaning the current version of the repository.  The value "latest"
will return the latest version of the CRAN repository.
If you have multiple repositories, you may also pass the R version number
of the desired repository.}

\item{libpaths}{A vector of paths specifying the locations of the repositories
to query. Default
is NULL, meaning the function will use the default R locations on the machine.
These default paths are identified by the Base R function \code{.libPaths()}.}
}
\value{
An object of class "pkgrepo".  The object will contain the
location of the repo, and a table of R packages with corresponding
version numbers.
}
\description{
Combines information related to a package repository. The function
retrieves all the packages in a repository for a specified version
of R.  Results will list the package name and version.
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_versions}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
