% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdiff.R
\docType{package}
\name{pkgdiff}
\alias{pkgdiff}
\alias{pkgdiff-package}
\title{Identify Differences Between Package Versions}
\description{
The \strong{pkgdiff} package contains functions to
identify differences between versions of R packages. The \strong{pkgdiff}
package was written specifically to detect breaking changes and
help with package upgrades.
}
\section{Functions}{

The main functions included in the \strong{pkgdiff} package are
as follows:
\itemize{
\item {\code{\link{pkg_diff}}: Compares two package versions, and
returns an object that contains information about the differences.
Information returned from this function is relatively high level.}
\item {\code{\link{pkg_stability}}: Compiles information on package
stability over time, and calculates a score based on this information.
This function is useful when making package selection choices.
to a vector.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://pkgdiff.r-sassy.org}
  \item \url{https://github.com/dbosak01/pkgdiff}
  \item Report bugs at \url{https://github.com/dbosak01/pkgdiff/issues}
}

}
\author{
\strong{Maintainer}: David Bosak \email{dbosak01@gmail.com}

}
\keyword{internal}
