% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{print.pcache}
\alias{print.pcache}
\title{Print a Package Cache Object}
\usage{
\method{print}{pcache}(x, ...)
}
\arguments{
\item{x}{The package cache to print.}

\item{...}{Follow-on parameters to the print function.}
}
\description{
Print a Package Cache Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{pkg_versions}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
