% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{print.pdiff}
\alias{print.pdiff}
\title{Print a Package Difference Object}
\usage{
\method{print}{pdiff}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The package difference to print}

\item{...}{Follow-on parameters to the print function}

\item{verbose}{Whether to print in summary or list-style.}
}
\description{
Print a Package Difference Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()}
}
\concept{pdiff}
