% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{print.pdiff}
\alias{print.pdiff}
\title{Print a Package Difference Object}
\usage{
\method{print}{pdiff}(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{The package difference to print}

\item{...}{Follow-on parameters to the print function}

\item{verbose}{Whether to print all added and removed functions
and function parameters.  Default is TRUE.  If FALSE, the function
will print removed functions and parameters, but only print
counts of added functions and parameters.}
}
\description{
Print a Package Difference Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_stability}()},
\code{\link{pkg_versions}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff_score}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
