% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{print.pdiff_score}
\alias{print.pdiff_score}
\title{Print a Package Difference Stability Score Object}
\usage{
\method{print}{pdiff_score}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The package difference to print}

\item{...}{Follow-on parameters to the print function}

\item{verbose}{Whether to print in summary or list-style.}
}
\description{
Print a Package Difference Stability Score Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_stability}()},
\code{\link{pkg_versions}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
