% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgInfo.R
\name{print.pinfo}
\alias{print.pinfo}
\title{Print a Package Info Object}
\usage{
\method{print}{pinfo}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The package info to print.}

\item{...}{Follow-on parameters to the print function.}

\item{verbose}{Whether to print in summary or list-style.}
}
\description{
Print a Package Info Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{print.pdiff}()}
}
\concept{pdiff}
