% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgrepo.R
\name{print.prepo}
\alias{print.prepo}
\title{Print a Package Repo Object}
\usage{
\method{print}{prepo}(x, ...)
}
\arguments{
\item{x}{The package repo to print.}

\item{...}{Follow-on parameters to the print function.}
}
\description{
Print a Package Repo Object
}
\seealso{
Other pdiff: 
\code{\link{pkg_cache}()},
\code{\link{pkg_diff}()},
\code{\link{pkg_info}()},
\code{\link{pkg_repo}()},
\code{\link{pkg_versions}()},
\code{\link{print.pcache}()},
\code{\link{print.pdiff}()},
\code{\link{print.pinfo}()}
}
\concept{pdiff}
