% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{print.rstability}
\alias{print.rstability}
\title{Print a Repo Stability Object}
\usage{
\method{print}{rstability}(x, ...)
}
\arguments{
\item{x}{The repo stability object to print.}

\item{...}{Follow-on parameters to the print function.}
}
\description{
Print a Repo Stability Object
}
\seealso{
Other prepo: 
\code{\link{pkg_repo}()},
\code{\link{print.prepo}()},
\code{\link{print.rbreak}()},
\code{\link{repo_breakages}()},
\code{\link{repo_stability}()}
}
\concept{prepo}
