% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{repo_breakages}
\alias{repo_breakages}
\title{Identify Breakages for a Repository}
\usage{
repo_breakages(v1pkgs = "current", v2pkgs = "latest")
}
\arguments{
\item{v1pkgs}{A data frame that identifies the source packages
and versions.}

\item{v2pkgs}{A data frame that identifies the target packages
and versions.  By default,}
}
\description{
The \code{repo_breakages} function generates a data frame
of breakage information of a list of packages or entire repository.
}
\seealso{
Other reports: 
\code{\link{repo_stability}()}
}
\concept{reports}
