% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{repo_breakages}
\alias{repo_breakages}
\title{Identify Breakages for a Repository}
\usage{
repo_breakages(r1 = "current", r2 = "latest")
}
\arguments{
\item{r1}{A data frame that identifies the source repository packages
and versions. The default value is "current", which means
the function will use the current versions of all package in
the current R repository.}

\item{r2}{A data frame that identifies the target repository packages
and versions. The default value is "latest", which means
the function will use the latest versions of all packages on CRAN.}
}
\description{
The \code{repo_breakages} function generates a data frame
of breakage information of a list of packages or entire repository.
}
\seealso{
Other reports: 
\code{\link{repo_stability}()}
}
\concept{reports}
