% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{repo_stability}
\alias{repo_stability}
\title{Generate Stability Scores}
\usage{
repo_stability(pkgs, releases = NULL, months = NULL)
}
\arguments{
\item{pkgs}{A data frame that identifies the source packages
and versions.}

\item{releases}{An integer number of releases to assess stability.}

\item{months}{An integer number of months from the current date
to assess stability.}
}
\value{
A data frame of information regarding the stability of the
packages. There will be one row per package.
}
\description{
The function generates stability scores
for a vector of packages. If passing a large number of packages,
be prepared for the
function to run for a considerable amount of time.
}
\seealso{
Other reports: 
\code{\link{repo_breakages}()}
}
\concept{reports}
