% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_stability}
\alias{report_stability}
\title{Generate a Report of Upgrade Breakages}
\usage{
report_stability(pkgs)
}
\arguments{
\item{pkgs}{A data frame that identifies the source packages
and versions.}
}
\description{
Generate a Report of Upgrade Breakages
}
\seealso{
Other reports: 
\code{\link{report_breakages}()}
}
\concept{reports}
